Move command buffer + encoders to their own file
