+++ src/rhi-shared-fwd.h
@@ -1,15 +1,32 @@
 #pragma once
 
 namespace rhi {
-class Device;
+
+class AccelerationStructure;
+class Buffer;
+class Texture;
+
+class CommandBuffer;
 class CommandEncoder;
+class ComputePassEncoder;
+class RayTracingPassEncoder;
+
 class CommandList;
+
+class Device;
+
 class ShaderObjectLayout;
 class ShaderProgram;
+
+class QueryPool;
 class Pipeline;
+class RenderPipeline;
+class ComputePipeline;
+class RayTracingPipeline;
 class ShaderObject;
 class RootShaderObject;
 struct ExtendedShaderObjectTypeList;
+class ExtendedShaderObjectTypeListObject;
 
 typedef uint32_t ShaderComponentID;
 } // namespace rhi
