+++ src/rhi-shared.h
@@ -8,10 +8,12 @@
 
 #include "core/common.h"
 #include "core/short_vector.h"
+#include "reference.h"
 
 #include "rhi-shared-fwd.h"
 
 #include "device.h"
+#include "command-buffer.h"
 
 #include <map>
 #include <set>
@@ -33,140 +35,6 @@ struct DescStructHeader
     DescStructHeader* next;
 };
 
-// We use a `BreakableReference` to avoid the cyclic reference situation in rhi implementation.
-// It is a common scenario where objects created from an `IDevice` implementation needs to hold
-// a strong reference to the device object that creates them. For example, a `Buffer` or a
-// `CommandQueue` needs to store a `m_device` member that points to the `IDevice`. At the same
-// time, the device implementation may also hold a reference to some of the objects it created
-// to represent the current device/binding state. Both parties would like to maintain a strong
-// reference to each other to achieve robustness against arbitrary ordering of destruction that
-// can be triggered by the user. However this creates cyclic reference situations that break
-// the `RefPtr` recyling mechanism. To solve this problem, we instead make each object reference
-// the device via a `BreakableReference<TDeviceImpl>` pointer. A breakable reference can be
-// turned into a weak reference via its `breakStrongReference()` call.
-// If we know there is a cyclic reference between an API object and the device/pool that creates it,
-// we can break the cycle when there is no longer any public references that come from `ComPtr`s to
-// the API object, by turning the reference to the device object from the API object to a weak
-// reference.
-// The following example illustrate how this mechanism works:
-// Suppose we have
-// ```
-// class DeviceImpl : IDevice { RefPtr<ShaderObject> m_currentObject; };
-// class ShaderObjectImpl : IShaderObject { BreakableReference<DeviceImpl> m_device; };
-// ```
-// And the user creates a device and a shader object, then somehow having the device reference
-// the shader object (this may not happen in actual implemetations, we just use it to illustrate
-// the situation):
-// ```
-// ComPtr<IDevice> device = createDevice();
-// ComPtr<ISomeResource> res = device->createResourceX(...);
-// device->m_currentResource = res;
-// ```
-// This setup is robust to any destruction ordering. If user releases reference to `device` first,
-// then the device object will not be freed yet, since there is still a strong reference to the device
-// implementation via `res->m_device`. Next when the user releases reference to `res`, the public
-// reference count to `res` via `ComPtr`s will go to 0, therefore triggering the call to
-// `res->m_device.breakStrongReference()`, releasing the remaining reference to device. This will cause
-// `device` to start destruction, which will release its strong reference to `res` during execution of
-// its destructor. Finally, this will triger the actual destruction of `res`.
-// On the other hand, if the user releases reference to `res` first, then the strong reference to `device`
-// will be broken immediately, but the actual destruction of `res` will not start. Next when the user
-// releases `device`, there will no longer be any other references to `device`, so the destruction of
-// `device` will start, causing the release of the internal reference to `res`, leading to its destruction.
-// Note that the above logic only works if it is known that there is a cyclic reference. If there are no
-// such cyclic reference, then it will be incorrect to break the strong reference to `IDevice` upon
-// public reference counter dropping to 0. This is because the actual destructor of `res` take place
-// after breaking the cycle, but if the resource's strong reference to the device is already the last reference,
-// turning that reference to weak reference will immediately trigger destruction of `device`, after which
-// we can no longer destruct `res` if the destructor needs `device`. Therefore we need to be careful
-// when using `BreakableReference`, and make sure we only call `breakStrongReference` only when it is known
-// that there is a cyclic reference. Luckily for all scenarios so far this is statically known.
-template<typename T>
-class BreakableReference
-{
-private:
-    RefPtr<T> m_strongPtr;
-    T* m_weakPtr = nullptr;
-
-public:
-    BreakableReference() = default;
-
-    BreakableReference(T* p) { *this = p; }
-
-    BreakableReference(const RefPtr<T>& p) { *this = p; }
-
-    void setWeakReference(T* p)
-    {
-        m_weakPtr = p;
-        m_strongPtr = nullptr;
-    }
-
-    T& operator*() const { return *get(); }
-
-    T* operator->() const { return get(); }
-
-    T* get() const { return m_weakPtr; }
-
-    operator T*() const { return get(); }
-
-    void operator=(RefPtr<T>& p)
-    {
-        m_strongPtr = p;
-        m_weakPtr = p.Ptr();
-    }
-
-    void operator=(T* p)
-    {
-        m_strongPtr = p;
-        m_weakPtr = p;
-    }
-
-    void breakStrongReference() { m_strongPtr = nullptr; }
-
-    void establishStrongReference() { m_strongPtr = m_weakPtr; }
-};
-
-// Helpers for returning an object implementation as COM pointer.
-template<typename TInterface, typename TImpl>
-void returnComPtr(TInterface** outInterface, TImpl* rawPtr)
-{
-    static_assert(!std::is_base_of<RefObject, TInterface>::value, "TInterface must be an interface type.");
-    rawPtr->addRef();
-    *outInterface = rawPtr;
-}
-
-template<typename TInterface, typename TImpl>
-void returnComPtr(TInterface** outInterface, const RefPtr<TImpl>& refPtr)
-{
-    static_assert(!std::is_base_of<RefObject, TInterface>::value, "TInterface must be an interface type.");
-    refPtr->addRef();
-    *outInterface = refPtr.Ptr();
-}
-
-template<typename TInterface, typename TImpl>
-void returnComPtr(TInterface** outInterface, ComPtr<TImpl>& comPtr)
-{
-    static_assert(!std::is_base_of<RefObject, TInterface>::value, "TInterface must be an interface type.");
-    *outInterface = comPtr.detach();
-}
-
-// Helpers for returning an object implementation as RefPtr.
-template<typename TDest, typename TImpl>
-void returnRefPtr(TDest** outPtr, RefPtr<TImpl>& refPtr)
-{
-    static_assert(std::is_base_of<RefObject, TDest>::value, "TDest must be a non-interface type.");
-    static_assert(std::is_base_of<RefObject, TImpl>::value, "TImpl must be a non-interface type.");
-    *outPtr = refPtr.Ptr();
-    refPtr->addReference();
-}
-
-template<typename TDest, typename TImpl>
-void returnRefPtrMove(TDest** outPtr, RefPtr<TImpl>& refPtr)
-{
-    static_assert(std::is_base_of<RefObject, TDest>::value, "TDest must be a non-interface type.");
-    static_assert(std::is_base_of<RefObject, TImpl>::value, "TImpl must be a non-interface type.");
-    *outPtr = refPtr.detach();
-}
 
 class Fence : public IFence, public ComObject
 {
@@ -632,9 +500,6 @@ public:
     virtual Result collectSpecializationArgs(ExtendedShaderObjectTypeList& args) override;
 };
 
-struct BindingData
-{};
-
 struct SpecializationKey
 {
     short_vector<ShaderComponentID> componentIDs;
@@ -734,313 +599,6 @@ public:
 
 public:
     QueryPoolDesc m_desc;
-};
-
-template<typename TDevice>
-class CommandQueue : public ICommandQueue, public ComObject
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_ALL
-    ICommandQueue* getInterface(const Guid& guid)
-    {
-        if (guid == ISlangUnknown::getTypeGuid() || guid == ICommandQueue::getTypeGuid())
-            return static_cast<ICommandQueue*>(this);
-        return nullptr;
-    }
-
-    virtual void comFree() override { breakStrongReferenceToDevice(); }
-
-public:
-    CommandQueue(TDevice* device, QueueType type)
-    {
-        m_device.setWeakReference(device);
-        m_type = type;
-    }
-
-    void breakStrongReferenceToDevice() { m_device.breakStrongReference(); }
-    void establishStrongReferenceToDevice() { m_device.establishStrongReference(); }
-
-    // ICommandQueue implementation
-    virtual SLANG_NO_THROW QueueType SLANG_MCALL getType() override { return m_type; }
-
-public:
-    BreakableReference<TDevice> m_device;
-    QueueType m_type;
-};
-
-class RenderPassEncoder : public IRenderPassEncoder
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_QUERY_INTERFACE
-    IRenderPassEncoder* getInterface(const Guid& guid);
-    virtual SLANG_NO_THROW uint32_t SLANG_MCALL addRef() override { return 1; }
-    virtual SLANG_NO_THROW uint32_t SLANG_MCALL release() override { return 1; }
-
-public:
-    CommandEncoder* m_commandEncoder;
-    ComPtr<IRenderPipeline> m_pipeline;
-    RefPtr<RootShaderObject> m_rootObject;
-    RenderState m_renderState;
-    /// Command list, nullptr if pass encoder is not active.
-    CommandList* m_commandList = nullptr;
-
-    RenderPassEncoder(CommandEncoder* commandEncoder);
-
-    void writeRenderState();
-
-    // IRenderPassEncoder implementation
-    virtual SLANG_NO_THROW IShaderObject* SLANG_MCALL bindPipeline(IRenderPipeline* pipeline) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL bindPipeline(IRenderPipeline* pipeline, IShaderObject* rootObject) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL setRenderState(const RenderState& state) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL draw(const DrawArguments& args) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL drawIndexed(const DrawArguments& args) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL drawIndirect(
-        uint32_t maxDrawCount,
-        IBuffer* argBuffer,
-        uint64_t argOffset,
-        IBuffer* countBuffer = nullptr,
-        uint64_t countOffset = 0
-    ) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL drawIndexedIndirect(
-        uint32_t maxDrawCount,
-        IBuffer* argBuffer,
-        uint64_t argOffset,
-        IBuffer* countBuffer = nullptr,
-        uint64_t countOffset = 0
-    ) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL drawMeshTasks(uint32_t x, uint32_t y, uint32_t z) override;
-
-    // IPassEncoder implementation
-    virtual SLANG_NO_THROW void SLANG_MCALL pushDebugGroup(const char* name, float rgbColor[3]) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL popDebugGroup() override;
-    virtual SLANG_NO_THROW void SLANG_MCALL insertDebugMarker(const char* name, float rgbColor[3]) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL end() override;
-};
-
-class ComputePassEncoder : public IComputePassEncoder
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_QUERY_INTERFACE
-    IComputePassEncoder* getInterface(const Guid& guid);
-    virtual SLANG_NO_THROW uint32_t SLANG_MCALL addRef() override { return 1; }
-    virtual SLANG_NO_THROW uint32_t SLANG_MCALL release() override { return 1; }
-
-public:
-    CommandEncoder* m_commandEncoder = nullptr;
-    ComPtr<IComputePipeline> m_pipeline;
-    RefPtr<RootShaderObject> m_rootObject;
-    /// Command list, nullptr if pass encoder is not active.
-    CommandList* m_commandList;
-
-    ComputePassEncoder(CommandEncoder* commandEncoder);
-
-    void writeComputeState();
-
-    // IComputePassEncoder implementation
-    virtual SLANG_NO_THROW IShaderObject* SLANG_MCALL bindPipeline(IComputePipeline* pipeline) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    bindPipeline(IComputePipeline* pipeline, IShaderObject* rootObject) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL dispatchCompute(uint32_t x, uint32_t y, uint32_t z) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL dispatchComputeIndirect(IBuffer* argBuffer, uint64_t offset) override;
-
-    // IPassEncoder implementation
-    virtual SLANG_NO_THROW void SLANG_MCALL pushDebugGroup(const char* name, float rgbColor[3]) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL popDebugGroup() override;
-    virtual SLANG_NO_THROW void SLANG_MCALL insertDebugMarker(const char* name, float rgbColor[3]) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL end() override;
-};
-
-class RayTracingPassEncoder : public IRayTracingPassEncoder
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_QUERY_INTERFACE
-    IRayTracingPassEncoder* getInterface(const Guid& guid);
-    virtual SLANG_NO_THROW uint32_t SLANG_MCALL addRef() override { return 1; }
-    virtual SLANG_NO_THROW uint32_t SLANG_MCALL release() override { return 1; }
-
-public:
-    CommandEncoder* m_commandEncoder = nullptr;
-    ComPtr<IRayTracingPipeline> m_pipeline;
-    ComPtr<IShaderTable> m_shaderTable;
-    RefPtr<RootShaderObject> m_rootObject;
-    /// Command list, nullptr if pass encoder is not active.
-    CommandList* m_commandList;
-
-    RayTracingPassEncoder(CommandEncoder* commandEncoder);
-
-    void writeRayTracingState();
-
-    // IRayTracingPassEncoder implementation
-    virtual SLANG_NO_THROW IShaderObject* SLANG_MCALL
-    bindPipeline(IRayTracingPipeline* pipeline, IShaderTable* shaderTable) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    bindPipeline(IRayTracingPipeline* pipeline, IShaderTable* shaderTable, IShaderObject* rootObject) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    dispatchRays(uint32_t rayGenShaderIndex, uint32_t width, uint32_t height, uint32_t depth) override;
-
-    // IPassEncoder implementation
-    virtual SLANG_NO_THROW void SLANG_MCALL pushDebugGroup(const char* name, float rgbColor[3]) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL popDebugGroup() override;
-    virtual SLANG_NO_THROW void SLANG_MCALL insertDebugMarker(const char* name, float rgbColor[3]) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL end() override;
-};
-
-class CommandEncoder : public ICommandEncoder, public ComObject
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_ALL
-    ICommandEncoder* getInterface(const Guid& guid);
-
-public:
-    // Current command list to write to. Must be set by the derived class.
-    CommandList* m_commandList = nullptr;
-
-    RenderPassEncoder m_renderPassEncoder;
-    ComputePassEncoder m_computePassEncoder;
-    RayTracingPassEncoder m_rayTracingPassEncoder;
-
-    // List of persisted pipeline specialization data.
-    // This is populated during command encoding and later used when asynchronously resolving pipelines.
-    std::vector<RefPtr<ExtendedShaderObjectTypeListObject>> m_pipelineSpecializationArgs;
-
-    CommandEncoder();
-
-    virtual Device* getDevice() = 0;
-    virtual Result getBindingData(RootShaderObject* rootObject, BindingData*& outBindingData) = 0;
-
-    Result getPipelineSpecializationArgs(
-        IPipeline* pipeline,
-        IShaderObject* object,
-        ExtendedShaderObjectTypeListObject*& outSpecializationArgs
-    );
-    Result resolvePipelines(Device* device);
-
-    // ICommandEncoder implementation
-    virtual SLANG_NO_THROW IRenderPassEncoder* SLANG_MCALL beginRenderPass(const RenderPassDesc& desc) override;
-    virtual SLANG_NO_THROW IComputePassEncoder* SLANG_MCALL beginComputePass() override;
-    virtual SLANG_NO_THROW IRayTracingPassEncoder* SLANG_MCALL beginRayTracingPass() override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    copyBuffer(IBuffer* dst, Offset dstOffset, IBuffer* src, Offset srcOffset, Size size) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL copyTexture(
-        ITexture* dst,
-        SubresourceRange dstSubresource,
-        Offset3D dstOffset,
-        ITexture* src,
-        SubresourceRange srcSubresource,
-        Offset3D srcOffset,
-        Extents extent
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL copyTextureToBuffer(
-        IBuffer* dst,
-        Offset dstOffset,
-        Size dstSize,
-        Size dstRowStride,
-        ITexture* src,
-        SubresourceRange srcSubresource,
-        Offset3D srcOffset,
-        Extents extent
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL uploadTextureData(
-        ITexture* dst,
-        SubresourceRange subresourceRange,
-        Offset3D offset,
-        Extents extent,
-        SubresourceData* subresourceData,
-        uint32_t subresourceDataCount
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    uploadBufferData(IBuffer* dst, Offset offset, Size size, void* data) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL clearBuffer(IBuffer* buffer, const BufferRange* range = nullptr) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL clearTexture(
-        ITexture* texture,
-        const ClearValue& clearValue = ClearValue(),
-        const SubresourceRange* subresourceRange = nullptr,
-        bool clearDepth = true,
-        bool clearStencil = true
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    resolveQuery(IQueryPool* queryPool, uint32_t index, uint32_t count, IBuffer* buffer, uint64_t offset) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL buildAccelerationStructure(
-        const AccelerationStructureBuildDesc& desc,
-        IAccelerationStructure* dst,
-        IAccelerationStructure* src,
-        BufferWithOffset scratchBuffer,
-        uint32_t propertyQueryCount,
-        AccelerationStructureQueryDesc* queryDescs
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL copyAccelerationStructure(
-        IAccelerationStructure* dst,
-        IAccelerationStructure* src,
-        AccelerationStructureCopyMode mode
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL queryAccelerationStructureProperties(
-        uint32_t accelerationStructureCount,
-        IAccelerationStructure** accelerationStructures,
-        uint32_t queryCount,
-        AccelerationStructureQueryDesc* queryDescs
-    ) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    serializeAccelerationStructure(BufferWithOffset dst, IAccelerationStructure* src) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    deserializeAccelerationStructure(IAccelerationStructure* dst, BufferWithOffset src) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    convertCooperativeVectorMatrix(const ConvertCooperativeVectorMatrixDesc* descs, uint32_t descCount) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL setBufferState(IBuffer* buffer, ResourceState state) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL
-    setTextureState(ITexture* texture, SubresourceRange subresourceRange, ResourceState state) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL pushDebugGroup(const char* name, float rgbColor[3]) override;
-    virtual SLANG_NO_THROW void SLANG_MCALL popDebugGroup() override;
-    virtual SLANG_NO_THROW void SLANG_MCALL insertDebugMarker(const char* name, float rgbColor[3]) override;
-
-    virtual SLANG_NO_THROW void SLANG_MCALL writeTimestamp(IQueryPool* queryPool, uint32_t queryIndex) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL finish(ICommandBuffer** outCommandBuffer) override;
-};
-
-class CommandBuffer : public ICommandBuffer, public ComObject
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_ALL
-    ICommandBuffer* getInterface(const Guid& guid);
-
-public:
-    CommandBuffer()
-        : m_commandList(m_allocator, m_trackedObjects)
-    {
-    }
-    virtual ~CommandBuffer() = default;
-
-    virtual Result reset()
-    {
-        m_commandList.reset();
-        m_allocator.reset();
-        m_trackedObjects.clear();
-        return SLANG_OK;
-    }
-
-    ArenaAllocator m_allocator;
-    CommandList m_commandList;
-    std::set<RefPtr<RefObject>> m_trackedObjects;
 };
 
 enum class PipelineType
